requirements.txt
langchain~=0.1.12
langchain-community~=0.0.28
langchainhub~=0.1.15
langchain-anthropic~=0.1.4
langchain-mistralai~=0.1.0
langchain-openai~=0.0.8
langchain-google-vertexai~=0.1.1
python-dotenv~=1.0.1
pandas~=2.2.1
tqdm~=4.66.2
evaluate~=0.4.1
rouge_score~=0.1.2
matplotlib~=3.8.3
pytest~=8.1.1
deepeval~=0.21.0
# OPTIONAL. For huggingface models.
transformers~=4.39.3
accelerate~=0.28.0